Re-organize and merge test files
