+++ coverage.out
@@ -1,0 +1,3 @@
+mode: set
+_/home/havoc/.projects/Go-PHP-Engine/php.go:10.36,12.2 1 1
+_/home/havoc/.projects/Go-PHP-Engine/php.go:14.56,16.2 1 1
